+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -167,7 +167,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-13T10:29:20.072Z",
+    "2026-01-13T14:58:54.165Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -207,7 +207,7 @@ export const CreateContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2026-01-13T10:29:20.072Z"))
+  updated: z.date().default(() => new Date("2026-01-13T14:58:54.165Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {