+++ packages/orq-rc/RELEASES.md
@@ -5799,3 +5799,13 @@ Based on:
 - [typescript v4.2.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.26 - packages/orq-rc
+
+## 2026-01-13 15:01:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.0 (2.794.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.27 - packages/orq-rc